refactor: Simplify App component by removing unnecessary div wrapper
Add a new Hero component to the website module, which displays a welcome message and a button. Update the Home page to include the Hero component. Also, create a Header component in the website module to display the header of the website.

This commit simplifies the App component by removing the unnecessary div wrapper around the AppRoutes component. This improves the code readability and reduces the nesting level.

Refactor the App component to return only the AppRoutes component directly.

New files added:
- src/modules/website/components/Hero.jsx
- src/modules/website/layouts/Header/index.jsx

Modified files:
- src/App.jsx
- src/modules/website/pages/Home/index.jsx
- src/modules/website/pages/index.jsx
- src/routes.jsx
