+++ src/App.jsx
@@ -2,11 +2,7 @@ import React from 'react';
 import AppRoutes from './routes';
 
 function App() {
-  return (
-    <div>
-      <AppRoutes />
-    </div>
-  );
+  return <AppRoutes />;
 }
 
 export default App;
