+++ src/modules/website/components/Hero.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Hero = () => {
+  return (
+    <section>
+      <h1>Welcome to MyCube Energy</h1>
+      <p>Your reliable partner in sustainable energy solutions.</p>
+      <button>Learn More</button>
+    </section>
+  );
+};
+
+export default Hero;
