+++ src/modules/website/layouts/Header/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div className='container text-blue-300'>Header</div>
+  )
+}
+
+export default Header
