+++ src/modules/website/pages/Home/index.jsx
@@ -1,7 +1,12 @@
-import React from 'react';
+import React from "react";
+import Hero from "../../components/Hero";
 
 const Home = () => {
-  return <h2>Home Page</h2>;
+  return (
+    <main>
+      <Hero />
+    </main>
+  );
 };
 
 export default Home;
