+++ src/modules/website/pages/index.jsx
@@ -1,0 +1,4 @@
+import Home from './Home';
+import About from './About';
+
+export { Home, About };
